backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
autopep8==2.3.2
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-limiter==0.1.6
filelock==3.20.3
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
packageurl-python==0.17.6
packaging==25.0
passlib==1.7.4
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
psycopg2==2.9.11
py-serializable==2.1.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
starlette==0.50.0
tomli==2.4.0
tomli_w==1.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
